refactor: move contexts to entities
