+++ src/app/index.ts
@@ -1,4 +1,3 @@
 export * from './app.tsx';
-export * from './contexts';
 export * from './protected-route';
 export * from './protected-home-route';
