+++ src/entities/contexts/index.ts
@@ -1,7 +1,6 @@
 export * from './shopListContext';
 export * from './userContext';
 export * from './cardsContext';
-export * from './cardContext';
 export * from './sortedCardsContext';
 export * from './messagesContext';
 export * from './loadingContext';
