+++ src/entities/index.ts
@@ -7,3 +7,4 @@ export * from './promo-card';
 export * from './card-share-popup';
 export * from './user-delete-popup';
 export * from './activate-email-popup';
+export * from './contexts';
