+++ src/widgets/home/index.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { ShopListContext } from '~/app/contexts';
+import { ShopListContext } from '~/entities';
 import { Box, Stack, Typography } from '@mui/material';
 import { AccentButton } from '~/shared/ui';
 import { PromoSlider } from '~/features';
